requirements.txt
pydantic-ai
aiogram
openai
python-dotenv
google-api-python-client
google-auth-oauthlib